Fix(mobile): [File naming structure] move from expo 52 to 53
