+++ apps/mobile/app/components/screen.tsx
@@ -167,7 +167,7 @@ function ScreenWithScrolling(props: ScreenProps) {
 		style
 	} = props as ScrollScreenProps;
 
-	const ref = useRef<ScrollView>();
+	const ref = useRef<ScrollView>(null);
 
 	const { scrollEnabled, onContentSizeChange, onLayout } = useAutoPreset(props as AutoScreenProps);
