+++ apps/mobile/app/components/task-size-popup.tsx
@@ -16,7 +16,7 @@ import { Feather, AntDesign, Ionicons } from '@expo/vector-icons';
 import { spacing, useAppTheme, typography } from '../theme';
 import { useTaskSizes } from '../services/hooks/features/useTaskSizes';
 import { ITaskSizeItem } from '../services/interfaces/ITaskSize';
-import { BadgedTaskSize } from './SizeIcon';
+import { BadgedTaskSize } from './size-icon';
 import { translate } from '../i18n';
 import { BlurView } from 'expo-blur';
 import TaskSizeForm from '../screens/authenticated/task-size-screen/components/task-size-form';
