+++ apps/mobile/app/components/task-status.tsx
@@ -6,7 +6,7 @@ import { AntDesign, Feather } from '@expo/vector-icons';
 import { ITeamTask } from '../services/interfaces/ITask';
 import { observer } from 'mobx-react-lite';
 import { useTeamTasks } from '../services/hooks/features/useTeamTasks';
-import TaskStatusPopup from './TaskStatusPopup';
+import TaskStatusPopup from './task-status-popup';
 import { typography, useAppTheme } from '../theme';
 import { translate } from '../i18n';
 import { useTaskStatusValue } from './status-type';
