+++ apps/mobile/app/screens/error-screen/error-boundary.tsx
@@ -1,5 +1,5 @@
 import React, { Component, ErrorInfo, ReactNode } from 'react';
-import { ErrorDetails } from './ErrorDetails';
+import { ErrorDetails } from './error-details';
 
 interface Props {
 	children: ReactNode;
