+++ apps/mobile/app/services/client/api/auth/login.ts
@@ -57,7 +57,7 @@ export async function login(params: ILoginDataAPI) {
 		 */
 	} else {
 		// generate a random password
-		const password = '123456' || generateToken(8);
+		const password = '123456';
 		const email = inviteResponse.data.email;
 		const names = inviteResponse.data.fullName.split(' ');
