+++ apps/mobile/app/utils/format-date.ts
@@ -1,15 +1,15 @@
 import { Locale, format, parseISO, intervalToDuration } from 'date-fns';
 import I18n from 'i18n-js';
 
-import ar from 'date-fns/locale/ar-SA';
-import ko from 'date-fns/locale/ko';
-import en from 'date-fns/locale/en-US';
+import { arSA } from 'date-fns/locale/ar-SA';
+import { ko } from 'date-fns/locale/ko';
+import { enUS } from 'date-fns/locale/en-US';
 
 type Options = Parameters<typeof format>[2];
 
 const getLocale = (): Locale => {
 	const locale = I18n.currentLocale().split('-')[0];
-	return (locale === 'ar' ? ar : locale === 'ko' ? ko : en) as any;
+	return (locale === 'ar' ? arSA : locale === 'ko' ? ko : enUS) as any;
 };
 
 export const formatDate = (date: string, dateFormat?: string, options?: Options) => {
