+++ apps/mobile/metro.config.js
@@ -3,5 +3,5 @@ const { getDefaultConfig } = require('expo/metro-config');
 
 /** @type {import('expo/metro-config').MetroConfig} */
 const config = getDefaultConfig(__dirname);
-
+config.resolver.unstable_conditionNames = ['browser', 'require', 'react-native'];
 module.exports = config;
