+++ apps/mobile/package.json
@@ -58,38 +58,42 @@
 	},
 	"dependencies": {
 		"@expo-google-fonts/space-grotesk": "^0.2.2",
-		"@expo/metro-runtime": "~4.0.0",
+		"@expo/metro-runtime": "~5.0.4",
 		"@gorhom/bottom-sheet": "^5.1.2",
-		"@react-native-async-storage/async-storage": "1.23.1",
+		"@react-native-async-storage/async-storage": "2.1.2",
 		"@react-navigation/bottom-tabs": "^6.5.7",
 		"@react-navigation/drawer": "^6.6.2",
 		"@react-navigation/native": "^6.0.2",
 		"@react-navigation/native-stack": "^6.0.2",
 		"@react-navigation/stack": "~6.3.16",
-		"@sentry/react-native": "~6.3.0",
-		"@shopify/flash-list": "1.7.3",
+		"@sentry/react-native": "~6.10.0",
+		"@shopify/flash-list": "1.7.6",
 		"@tanstack/react-query": "^5.36.0",
 		"apisauce": "3.0.1",
+		"assert": "^2.1.0",
+		"browserify-zlib": "^0.2.0",
 		"date-fns": "^4.1.0",
-		"expo": "~52.0.7",
-		"expo-application": "~6.0.2",
-		"expo-blur": "^14.0.3",
-		"expo-build-properties": "~0.13.1",
-		"expo-clipboard": "~7.0.1",
-		"expo-constants": "~17.0.8",
-		"expo-dev-client": "~5.0.1",
-		"expo-device": "~7.0.2",
-		"expo-file-system": "^18.0.11",
-		"expo-font": "~13.0.1",
-		"expo-image-picker": "~16.0.6",
-		"expo-linear-gradient": "~14.0.2",
-		"expo-linking": "~7.0.2",
-		"expo-localization": "~16.0.0",
-		"expo-media-library": "~17.0.6",
-		"expo-splash-screen": "~0.29.10",
-		"expo-status-bar": "~2.0.0",
-		"expo-system-ui": "~4.0.3",
-		"expo-updates": "~0.27.4",
+		"expo": "^53.0.0",
+		"expo-application": "~6.1.4",
+		"expo-blur": "~14.1.4",
+		"expo-build-properties": "~0.14.6",
+		"expo-clipboard": "~7.1.4",
+		"expo-constants": "~17.1.5",
+		"expo-crypto": "^14.1.4",
+		"expo-dev-client": "~5.1.8",
+		"expo-device": "~7.1.4",
+		"expo-file-system": "~18.1.9",
+		"expo-font": "~13.3.1",
+		"expo-image-picker": "~16.1.4",
+		"expo-linear-gradient": "~14.1.4",
+		"expo-linking": "~7.1.4",
+		"expo-localization": "~16.1.5",
+		"expo-media-library": "~17.1.6",
+		"expo-splash-screen": "~0.30.8",
+		"expo-status-bar": "~2.2.3",
+		"expo-system-ui": "~5.0.7",
+		"expo-updates": "~0.28.12",
+		"https-browserify": "^1.0.0",
 		"i18n-js": "3.8",
 		"i18next": "^23.14.0",
 		"intl-pluralrules": "^2.0.1",
@@ -100,40 +104,49 @@
 		"mobx-react-lite": "4.0.5",
 		"mobx-state-tree": "5.3.0",
 		"moment-timezone": "^0.5.42",
-		"react": "18.3.1",
-		"react-dom": "18.3.1",
+		"path-browserify": "^1.0.1",
+		"react": "19.0.0",
+		"react-dom": "19.0.0",
 		"react-i18next": "^15.0.1",
-		"react-native": "0.76.7",
+		"react-native": "0.79.2",
 		"react-native-animatable": "^1.3.3",
-		"react-native-bootsplash": "4.5.3",
+		"react-native-bootsplash": "^6.3.7",
 		"react-native-calendars": "^1.1302.0",
 		"react-native-circular-progress": "^1.3.8",
 		"react-native-click-outside": "^0.1.1",
 		"react-native-cn-quill": "^0.7.18",
 		"react-native-confirmation-code-field": "^7.3.2",
+		"react-native-crypto": "^2.2.0",
 		"react-native-dotenv": "^3.4.11",
 		"react-native-drawer-layout": "^4.0.1",
 		"react-native-dropdown-picker": "^5.4.6",
 		"react-native-extended-stylesheet": "^0.12.0",
 		"react-native-flash-message": "^0.4.0",
-		"react-native-gesture-handler": "~2.20.2",
+		"react-native-fs": "^2.20.0",
+		"react-native-gesture-handler": "~2.24.0",
 		"react-native-keyboard-controller": "^1.12.7",
 		"react-native-mmkv": "^2.12.2",
-		"react-native-pager-view": "6.5.1",
+		"react-native-pager-view": "6.7.1",
 		"react-native-paper": "5.5.1",
-		"react-native-reanimated": "~3.16.1",
-		"react-native-safe-area-context": "4.12.0",
-		"react-native-screens": "~4.4.0",
+		"react-native-reanimated": "~3.17.4",
+		"react-native-safe-area-context": "5.4.0",
+		"react-native-screens": "~4.10.0",
 		"react-native-skeletons": "^1.3.1",
-		"react-native-svg": "15.8.0",
+		"react-native-svg": "15.11.2",
 		"react-native-svg-transformer": "^1.5.0",
 		"react-native-tab-view": "^3.5.2",
+		"react-native-tcp": "^4.0.0",
 		"react-native-vector-icons": "^9.2.0",
-		"react-native-web": "~0.19.6",
-		"react-native-webview": "13.12.5",
+		"react-native-web": "^0.20.0",
+		"react-native-webview": "13.13.5",
+		"readable-stream": "^4.7.0",
 		"reanimated-color-picker": "^2.3.4",
-		"sentry-expo": "~7.0.0",
-		"validator": "^13.11.0"
+		"sentry-expo": "^7.2.0",
+		"stream-browserify": "^3.0.0",
+		"stream-http": "^3.2.0",
+		"tslib": "^2.6.0",
+		"validator": "^13.11.0",
+		"whatwg-url": "^14.2.0"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.20.0",
@@ -141,17 +154,18 @@
 		"@babel/runtime": "^7.20.0",
 		"@hapi/shot": "^6.0.1",
 		"@testing-library/react-native": "^12.5.2",
+		"@types/hapi__shot": "^6.0.0",
 		"@types/jest": "^29.5.14",
-		"@types/react": "~18.3.12",
+		"@types/react": "~19.0.10",
 		"babel-jest": "^29.2.1",
 		"eslint": "^8.57.0",
-		"eslint-config-expo": "~8.0.1",
+		"eslint-config-expo": "~9.2.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-prettier": "^5.2.1",
 		"eslint-plugin-react-native": "^4.1.0",
 		"eslint-plugin-reactotron": "^0.1.2",
 		"jest": "^29.2.1",
-		"jest-expo": "~52.0.1",
+		"jest-expo": "~53.0.4",
 		"patch-package": "^8.0.0",
 		"postinstall-prepare": "1.0.1",
 		"prettier": "^3.3.3",
@@ -163,8 +177,7 @@
 		"reactotron-react-native-mmkv": "^0.2.6",
 		"ts-jest": "^29.1.1",
 		"ts-node": "^10.9.2",
-		"@types/hapi__shot": "^6.0.0",
-		"typescript": "~5.3.3"
+		"typescript": "~5.8.3"
 	},
 	"prettier": {
 		"printWidth": 120,
